Edit : users app/urls.py and views.py(2)
